mauri web
Mauri web is a Python script designed to simulate a "flood" attack, specifically a TCP flood, on a specified server. The script utilizes multiple threads to establish concurrent connections and inundate the target server with traffic. The goal of this tool is to test the server's resilience to such flood attacks and analyze its performance under stress.